[Format]

+++ util/time.go
@@ -4,5 +4,5 @@ import "time"
 
 func CurrMillSecond() int64 {
 	now := time.Now()
-	return int64(now.UnixNano()) / int64(time.Millisecond/time.Nanosecond)
+	return now.UnixNano() / int64(time.Millisecond/time.Nanosecond)
 }
